+++ ChargingApp/Entity/VIPLevels.cs
@@ -3,6 +3,6 @@
 public class VIPLevel :BaseEntity
 {
     public int VipLevel { get; set; }
-    public int BenefitPercent { get; set; }
-    public int MinimumPurchase { get; set; } = 0;
+    public double BenefitPercent { get; set; }
+    public double MinimumPurchase { get; set; } = 0;
 }
